Refactor import paths and adjust password reset parameters order
