+++ src/pages/auth/changePassword.tsx
@@ -2,11 +2,10 @@ import { Loader2 } from 'lucide-react';
 import { Fragment } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 
-import AlertError from '../../common/components/Form/AlertError';
-import ErrorMessage from '../../common/components/Form/ErrorMessage';
-import PasswordInput from '../../common/components/Input/PasswordInput';
-import { Button } from '../../common/components/ui/button';
-
+import AlertError from '@/common/components/form/alertError';
+import ErrorMessage from '@/common/components/form/errorMessage';
+import PasswordInput from '@/common/components/input/passwordInput';
+import { Button } from '@/common/components/ui/button';
 import { useAuthProvider } from '@/modules/auth/hooks/useAuthProvider';
 import { AUTH_VALIDATIONS } from '@/modules/auth/message/auth-messages';
