+++ src/pages/auth/protectedRoute.tsx
@@ -1,7 +1,7 @@
 import { Fragment, ReactElement, ReactNode, useEffect } from 'react';
 import { Navigate } from 'react-router-dom';
 
-import FullscreenLoading from '@/common/views/FullscreenLoading';
+import FullscreenLoading from '@/common/views/fullscreenLoading';
 import { useAuthProvider } from '@/modules/auth/hooks/useAuthProvider';
 
 function ProtectedRoute({
