+++ src/pages/auth/recoverPassword.tsx
@@ -3,10 +3,9 @@ import { Fragment } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { Link } from 'react-router-dom';
 
-import ErrorMessage from '../../common/components/Form/ErrorMessage';
-import { Button } from '../../common/components/ui/button';
-import { Input } from '../../common/components/ui/input';
-
+import ErrorMessage from '@/common/components/form/errorMessage';
+import { Button } from '@/common/components/ui/button';
+import { Input } from '@/common/components/ui/input';
 import { useAuthProvider } from '@/modules/auth/hooks/useAuthProvider';
 import { AUTH_VALIDATIONS } from '@/modules/auth/message/auth-messages';
